+++ volmdlr/faces.py
@@ -21,11 +21,9 @@ import volmdlr.display as vmd
 import volmdlr.edges as vme
 import volmdlr.geometry
 import volmdlr.grid
-import volmdlr.utils.parametric as vm_parametric
 from volmdlr import surfaces
 import volmdlr.wires
-import os
-c = 0
+
 
 class Face3D(volmdlr.core.Primitive3D):
     """
@@ -172,11 +170,7 @@ class Face3D(volmdlr.core.Primitive3D):
         outer_contour2d = None
         outer_contour3d, inner_contours3d = None, None
         if lc3d == 1:
-            try:
-                outer_contour2d = surface.contour3d_to_2d(contours3d[0])
-            except Exception:
-                surface.save_to_file("contour3d_to_2d_surface.json")
-                contours3d[0].save_to_file("contour3d_to_2d_contour.json")
+            outer_contour2d = surface.contour3d_to_2d(contours3d[0])
             outer_contour3d = contours3d[0]
             inner_contours2d = []
 
@@ -209,17 +203,7 @@ class Face3D(volmdlr.core.Primitive3D):
 
         # if outer_contour3d and not outer_contour3d.is_ordered():
         #     outer_contour2d = vm_parametric.contour2d_healing(outer_contour2d)
-            # if len(outer_contour3d.primitives) == 2:
-            #     if any(primitive is None for primitive in outer_contour2d.primitives):
-            #         surface.save_to_file("contour2d_healing_bug_surface.json")
-            #         outer_contour3d.save_to_file("contour2d_healing_bug_contour.json")
-            #     ax = outer_contour2d.plot()
-            #     ax.set_aspect("auto")
         if (not outer_contour2d) or (not outer_contour2d.primitives):
-            # onlyfiles = next(os.walk("none/surfaces"))[2]  # directory is your directory path as string
-            # i = len(onlyfiles)
-            # surface.save_to_file(f"none/surfaces/none_contour3d_to_2d_surface_{i}.json")
-            # contours3d[0].save_to_file(f"none/contours/none_contour3d_to_2d_contour{i}.json")
             return None
         surface2d = surfaces.Surface2D(outer_contour=outer_contour2d,
                                        inner_contours=inner_contours2d)
@@ -2793,9 +2777,9 @@ class BSplineFace3D(Face3D):
         Specifies an adapted size of the discretization grid used in face triangulation.
         """
         if self.surface3d.x_periodicity or self.surface3d.y_periodicity:
-            resolution = 25
+            resolution = 10
         else:
-            resolution = 15
+            resolution = 10
         u_min, u_max, v_min, v_max = self.surface2d.bounding_rectangle().bounds()
         delta_u = u_max - u_min
         number_points_x = int(delta_u * resolution)