+++ volmdlr/shells.py
@@ -1,4 +1,5 @@
 """volmdlr shells module."""
+import traceback
 import warnings
 from itertools import chain
 from typing import List, Tuple
@@ -585,13 +586,17 @@ class OpenShell3D(volmdlr.core.CompositePrimitive3D):
         for i, face in enumerate(self.faces):
             try:
                 face_mesh = face.triangulation()
+
             except Exception:
                 warnings.warn(f"Could not triangulate {face.__class__.__name__} with index {i} in the shell "
                               f"{self.name} faces. Probabaly because topology error in contour2d.")
+                print(traceback.format_exc())
                 continue
-            if not face_mesh:
-                face.save_to_file("face_triangulation_none.json")
-            meshes.append(face_mesh)
+            if face_mesh:
+                meshes.append(face_mesh)
+            else:
+                warnings.warn(f"Could not triangulate {face.__class__.__name__} with index {i} in the shell "
+                              f"{self.name} faces. Probabaly because topology error in contour2d.")
         return display.DisplayMesh3D.merge_meshes(meshes)
 
     def babylon_meshes(self, merge_meshes=True):