+++ volmdlr/wires.py
@@ -33,7 +33,6 @@ import volmdlr.utils.intersections as vm_utils_intersections
 from volmdlr.core_compiled import polygon_point_belongs
 from volmdlr.core import EdgeStyle
 
-
 def argmax(list_of_numbers):
     """
     Returns the max value and the argmax.